Sat, Jan 21, 2023, 12:33 PM +00:00

+++ test/learn.c
@@ -2,8 +2,8 @@
 
 int main() {
 
-    int n, a, b, c, d, e, f, g;
-    scanf("%d",&n);
+    long long int n, a, b, c, d, e, f, g;
+    scanf("%lld",&n);
     a = n/100;
     b = (n- a*100)/50;
     c = (n- a*100 - b*50)/20;
@@ -12,13 +12,13 @@ int main() {
     f = (n - a*100 - b*50 - c*20 -d*10 - e*5)/2;
     g= (n - a*100 - b*50 - c*20 -d*10 - e*5 - f*2)/1;
 
-    printf("%d nota(s) de R$ 100,00 \n",a);
-    printf("%d nota(s) de R$ 50,00 \n",b);
-    printf("%d nota(s) de R$ 20,00 \n",c);
-    printf("%d nota(s) de R$ 10,00 \n",d);
-    printf("%d nota(s) de R$ 5,00 \n",e);
-    printf("%d nota(s) de R$ 2,00 \n",f);
-    printf("%d nota(s) de R$ 1,00 \n",g);
+    printf("%lld nota(s) de R$ 100,00 \n",a);
+    printf("%lld nota(s) de R$ 50,00 \n",b);
+    printf("%lld nota(s) de R$ 20,00 \n",c);
+    printf("%lld nota(s) de R$ 10,00 \n",d);
+    printf("%lld nota(s) de R$ 5,00 \n",e);
+    printf("%lld nota(s) de R$ 2,00 \n",f);
+    printf("%lld nota(s) de R$ 1,00 \n",g);
 
     return 0;
 }
